Audit 1-4: 4 audits + 12 fixes applied
4 sequential audits conducted (research only, no code changes):
- AUDIT-1: Core layer (89 files, 33 issues: 3 critical, 11 major, 19 minor)
- AUDIT-2: Modules layer (95 files, 48 issues: 4 critical, 14 major, 30 minor)
- AUDIT-3: Entry + Adapter (34 files, 47 issues: 5 critical, 14 major, 28 minor)
- AUDIT-4: Docs compliance (25 docs, 2 P0 formula mismatches, 8 auto-fixable,
  4 conceptual differences requiring user decision)

12 fixes applied (auto-fixable + critical, docs canonical):

1. qiCost formula: capacity*0.15 -> floor(baseCapacity × 2^(level-1))
   (TechniqueGeneratorService.cs — TECHNIQUE_SYSTEM.md §5.2)

2. UltimateQiCostMultiplier: 1.5f -> 2.0f + _PERMIL 1500 -> 2000
   (Constants.cs — TECHNIQUE_SYSTEM.md §9.1)

3. TechniqueData comments corrected to match doc formulas

4. StorageRingService: float qiCost -> long qiCost (ЗАПРЕТ 2 — no float Qi)

5. GameLifetimeScope: Charger moved position 6 -> 14 (DI_AND_EVENTBUS §1.2)

6. Esc-close inventory now resumes Time (was leaving game paused)

7. HandlePickup distance: 1.5f*96f -> 1.5f*GameConstants.TILE_PIXELS

8. GroundItemRenderer ZIndex: Objects+1 -> Objects (items below player)

9. PlayerInputService: IsSticky('minimap') -> IsSticky('meditate') (copy-paste bug)

10. InputMapInitializer: added 'meditate' action (Key.V)

11. PlayerInputService: removed dead IsSticky('i') check

12. InventoryWindow: TestItemSeeder gated behind #if DEBUG

Build: 0 errors, 227 warnings (pre-existing).
Headless: all 16 modules start in canonical order.

16 conceptual differences remain for user decision (documented in
checkpoints/08_21_audit_fix_plan.md).

+++ game/src/Core/Interfaces/IStorageRingService.cs
@@ -18,19 +18,19 @@ namespace CultivationGame.Core.Interfaces
     public interface IStorageRingService
     {
         /// <summary>Поместить предмет в хранилище кольца</summary>
-        bool TryStore(string ringItemId, ItemData item, out float qiCost);
+        bool TryStore(string ringItemId, ItemData item, out long qiCost);
 
         /// <summary>Извлечь предмет из хранилища кольца</summary>
-        bool TryRetrieve(string ringItemId, string storedItemId, out ItemData item, out float qiCost);
+        bool TryRetrieve(string ringItemId, string storedItemId, out ItemData item, out long qiCost);
 
         /// <summary>Получить содержимое хранилища кольца</summary>
         IReadOnlyList<InventorySlot> GetRingContents(string ringItemId);
 
         /// <summary>Qi стоимость помещения предмета</summary>
-        float GetStoreQiCost(int ringTier, float itemWeight);
+        long GetStoreQiCost(int ringTier, float itemWeight);
 
         /// <summary>Qi стоимость извлечения предмета</summary>
-        float GetRetrieveQiCost(int ringTier, float itemWeight);
+        long GetRetrieveQiCost(int ringTier, float itemWeight);
 
         /// <summary>Активировать хранилище для кольца (при экипировке)</summary>
         void ActivateRing(string ringItemId, int tier, int capacity);
